Se limpiaron servicios

+++ src/main/java/com/backend/digitalhouse/ClinicaOdontologica/dto/salida/domicilio/DomicilioSalidaDto.java
@@ -1,7 +1,7 @@
 package com.backend.digitalhouse.ClinicaOdontologica.dto.salida.domicilio;
 
 public class DomicilioSalidaDto {
-    private int id;
+    private Long id;
     private String calle;
     private int numero;
     private String localidad;
@@ -9,7 +9,7 @@ public class DomicilioSalidaDto {
 
     public DomicilioSalidaDto(){}
 
-    public DomicilioSalidaDto(int id, String calle, int numero, String localidad, String provincia) {
+    public DomicilioSalidaDto(Long id, String calle, int numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -17,9 +17,9 @@ public class DomicilioSalidaDto {
         this.provincia = provincia;
     }
 
-    public int getId(){return id;}
+    public Long getId(){return id;}
 
-    public void setId(int id){this.id = id;}
+    public void setId(Long id){this.id = id;}
 
     public String getCalle() {
         return calle;
