+++ src/main/java/com/backend/digitalhouse/ClinicaOdontologica/dto/salida/paciente/PacienteSalidaDto.java
@@ -5,7 +5,7 @@ import com.backend.digitalhouse.ClinicaOdontologica.dto.salida.domicilio.Domicil
 import java.time.LocalDate;
 
 public class PacienteSalidaDto {
-    private int id;
+    private Long id;
     private String nombre;
 
     private String apellido;
@@ -17,7 +17,7 @@ public class PacienteSalidaDto {
     public PacienteSalidaDto() {
     }
 
-    public PacienteSalidaDto(int id, String nombre, String apellido, String dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilio) {
+    public PacienteSalidaDto(Long id, String nombre, String apellido, String dni, LocalDate fechaIngreso, DomicilioSalidaDto domicilio) {
         this.id = id;
         this.nombre = nombre;
         this.apellido = apellido;
@@ -26,11 +26,11 @@ public class PacienteSalidaDto {
         this.domicilio = domicilio;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
